renderercommon: fix warning

+++ src/renderercommon/tr_public.h
@@ -186,7 +186,7 @@ typedef struct
 #ifdef HUNK_DEBUG
 	void * (*Hunk_AllocDebug)(unsigned int size, ha_pref pref, char *label, char *file, int line);
 #else
-	void * (*Hunk_Alloc)(size_t size, ha_pref pref);
+	void * (*Hunk_Alloc)(unsigned int size, ha_pref pref);
 #endif
 	void * (*Hunk_AllocateTempMemory)(unsigned int size);
 	void (*Hunk_FreeTempMemory)(void *block);
